Move JSON related classes to appropriate package CIRC-381
