+++ src/main/java/org/folio/circulation/domain/LoanRepresentation.java
@@ -6,7 +6,7 @@ import static org.folio.circulation.domain.representations.LoanProperties.LOAN_P
 import static org.folio.circulation.domain.representations.LoanProperties.LOST_ITEM_POLICY;
 import static org.folio.circulation.domain.representations.LoanProperties.OVERDUE_FINE_POLICY;
 import static org.folio.circulation.domain.representations.LoanProperties.PATRON_GROUP_ID_AT_CHECKOUT;
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import java.lang.invoke.MethodHandles;
