+++ src/main/java/org/folio/circulation/domain/representations/CallNumberComponentsRepresentation.java
@@ -1,6 +1,6 @@
 package org.folio.circulation.domain.representations;
 
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import org.folio.circulation.domain.CallNumberComponents;
