+++ src/main/java/org/folio/circulation/infrastructure/storage/inventory/ItemRepository.java
@@ -3,7 +3,7 @@ package org.folio.circulation.infrastructure.storage.inventory;
 import static java.util.Objects.isNull;
 import static java.util.concurrent.CompletableFuture.completedFuture;
 import static java.util.function.Function.identity;
-import static org.folio.circulation.support.JsonKeys.byId;
+import static org.folio.circulation.support.json.JsonKeys.byId;
 import static org.folio.circulation.support.results.Result.ofAsync;
 import static org.folio.circulation.support.results.Result.succeeded;
 import static org.folio.circulation.support.results.ResultBinding.flatMapResult;
@@ -325,7 +325,7 @@ public class ItemRepository {
       }
     });
   }
-  
+
   public <T extends ItemRelatedRecord> CompletableFuture<Result<MultipleRecords<T>>> fetchItemsFor(
     Result<MultipleRecords<T>> result,
     BiFunction<T, Item, T> includeItemMap) {
