+++ src/test/java/api/support/builders/DeclareClaimedReturnedItemAsMissingRequestBuilder.java
@@ -1,6 +1,6 @@
 package api.support.builders;
 
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import io.vertx.core.json.JsonObject;
