+++ src/test/java/api/support/builders/FeeFineOwnerBuilder.java
@@ -1,6 +1,6 @@
 package api.support.builders;
 
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import java.util.Arrays;
 import java.util.List;
