+++ src/test/java/api/support/builders/InstanceBuilder.java
@@ -1,6 +1,6 @@
 package api.support.builders;
 
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import java.util.ArrayList;
 import java.util.Collections;
