+++ src/test/java/api/support/fixtures/AddressTypesFixture.java
@@ -3,7 +3,7 @@ package api.support.fixtures;
 import static api.support.fixtures.AddressExamples.HOME_ADDRESS_TYPE;
 import static api.support.fixtures.AddressExamples.WORK_ADDRESS_TYPE;
 import static org.folio.circulation.support.json.JsonPropertyFetcher.getProperty;
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import org.folio.circulation.support.http.client.IndividualResource;
