+++ src/test/java/api/support/fixtures/LocationsFixture.java
@@ -1,7 +1,7 @@
 package api.support.fixtures;
 
 import static org.folio.circulation.support.json.JsonPropertyFetcher.getProperty;
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 
 import java.util.function.Function;
