+++ src/test/java/org/folio/circulation/domain/ConfigurationServiceTest.java
@@ -1,6 +1,6 @@
 package org.folio.circulation.domain;
 
-import static org.folio.circulation.support.JsonPropertyWriter.write;
+import static org.folio.circulation.support.json.JsonPropertyWriter.write;
 import static org.junit.Assert.assertEquals;
 
 import java.util.Collections;
